src/main/java/Passanger.java
public class Passanger {

    private String name;
    private int bags;
    private Flight flight;
    private int seatNumber;

    public Passanger(String name, int bags) {
        this.name = name;
        this.bags = bags;


    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getBags() {
        return bags;
    }

    public void setBags(int bags) {
        this.bags = bags;
    }

    public void setFlight(Flight flight) {
        this.flight = flight;
    }

    public Flight getFlight() {
        return flight;
    }

    public int getSeatNumber() {
        return seatNumber;
    }

    public void setSeatNumber(int seatNumber) {
        this.seatNumber = seatNumber;
    }


}
